Add: 7) Aralash: a + b / c (2 kasr) [otkirovasrorbek450-boop]

+++ lms_kod_topshiriqlari/01-Asoslar/09-Arifmetik_operatorlar_____-_____________va_priorit/07-7__Aralash__a___b___c__2_kasr_.py
@@ -1,3 +1,5 @@
-a, b, c = map(float, input().split())
+a, b, c = map(int, input().split())
+
 natija = a + (b / c)
+
 print(f"Result: {natija:.2f}")
